+++ image-recognition/src/main/java/by/sunnycore/recognition/image/cluster/DistanceCounter.java
@@ -2,5 +2,5 @@ package by.sunnycore.recognition.image.cluster;
 
 public interface DistanceCounter {
 
-	double countDistance(double[] dot1,double[] dot2);
+	double countDistance(short[] dot1,short[] dot2);
 }
