+++ image-recognition/src/main/java/by/sunnycore/recognition/image/cluster/impl/EuklidDistanceCounter.java
@@ -10,7 +10,7 @@ import by.sunnycore.recognition.image.cluster.DistanceCounter;
 public class EuklidDistanceCounter implements DistanceCounter{
 
 	@Override
-	public double countDistance(double[] dot1, double[] dot2) {
+	public double countDistance(short[] dot1, short[] dot2) {
 		double sum = 0;
 		for(int i=0;i<dot1.length;i++){
 			sum+=Math.pow((dot1[i]-dot2[i]), 2);
